# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface Query {
  id: number;
  name: string;
  profile: string;
  mobile: string;
  email: string;
  queryText: string;
  replied: boolean;
  createdAt: string;
}

export interface QueryInput {
  /** @minLength 1 */
  name: string;
  /** @minLength 1 */
  profile: string;
  /** @minLength 1 */
  mobile: string;
  email: string;
  /** @minLength 10 */
  queryText: string;
}

export interface AdminLoginInput {
  password: string;
}

export interface AdminToken {
  token: string;
}

export interface ErrorResponse {
  error: string;
}