# lib/api-zod/src/generated/types/query.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface Query {
  id: number;
  name: string;
  profile: string;
  mobile: string;
  email: string;
  queryText: string;
  replied: boolean;
  createdAt: string;
}