# lib/api-zod/src/generated/types/queryInput.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface QueryInput {
  /** @minLength 1 */
  name: string;
  /** @minLength 1 */
  profile: string;
  /** @minLength 1 */
  mobile: string;
  email: string;
  /** @minLength 10 */
  queryText: string;
}